Refactor SQLDBCheck hasReplicasConfigured method for improved accuracy
The commit modifies the `hasReplicasConfigured` method in the `SQLDBCheck` class to improve the accuracy of checking if there are replicas configured for the database. Instead of checking if the result of `az sql db replica list-links` is greater than 0, it now checks if the count of the result is greater than 0. This ensures that the method correctly identifies if there are replicas configured.